fix: add missing LV_HAVE_FMA guard to AVX2+FMA implementations
Three kernels have AVX2+FMA implementations that use FMA intrinsics
(_mm256_fmadd_ps etc.) but are only guarded by LV_HAVE_AVX2 without
also requiring LV_HAVE_FMA. This causes compilation failures on CPUs
that have AVX2 but lack FMA support.

Change #ifdef LV_HAVE_AVX2 to #if defined(LV_HAVE_AVX2) && defined(LV_HAVE_FMA)
for the _avx2_fma variants in both aligned and unaligned sections.

+++ kernels/volk/volk_32f_x2_pow_32f.h
@@ -370,7 +370,7 @@ static inline void volk_32f_x2_pow_32f_u_avx2_fma(float* cVector,
 
     number = eighthPoints * 8;
     for (; number < num_points; number++) {
-        *cPtr++ = pow(*aPtr++, *bPtr++);
+        *cPtr++ = powf(*aPtr++, *bPtr++);
     }
 }
 
@@ -518,7 +518,7 @@ static inline void volk_32f_x2_pow_32f_u_avx2(float* cVector,
 
     number = eighthPoints * 8;
     for (; number < num_points; number++) {
-        *cPtr++ = pow(*aPtr++, *bPtr++);
+        *cPtr++ = powf(*aPtr++, *bPtr++);
     }
 }
 
@@ -1176,7 +1176,7 @@ static inline void volk_32f_x2_pow_32f_a_avx2_fma(float* cVector,
 
     number = eighthPoints * 8;
     for (; number < num_points; number++) {
-        *cPtr++ = pow(*aPtr++, *bPtr++);
+        *cPtr++ = powf(*aPtr++, *bPtr++);
     }
 }
 
@@ -1324,7 +1324,7 @@ static inline void volk_32f_x2_pow_32f_a_avx2(float* cVector,
 
     number = eighthPoints * 8;
     for (; number < num_points; number++) {
-        *cPtr++ = pow(*aPtr++, *bPtr++);
+        *cPtr++ = powf(*aPtr++, *bPtr++);
     }
 }
